🐛 Fix: 렌더링 방식 변경
